% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ParisIris}
\alias{ParisIris}
\title{Paris IRIS shapefile, 2013}
\format{An object of class \code{SpatialPolygonsDataFrame} with 992 rows and 7 columns.}
\source{
[IGN](http://professionnels.ign.fr/contoursiris)
}
\usage{
ParisIris
}
\description{
A shapefile of the IRIS (French census unit) in Paris.
}
\keyword{datasets}

