% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ParisPollingStations2012}
\alias{ParisPollingStations2012}
\title{Paris polling stations shapefile, 2012}
\format{An object of class \code{SpatialPolygonsDataFrame} with 867 rows and 5 columns.}
\source{
[Paris city](http://opendata.paris.fr/explore/dataset/zones-de-rattachement-des-bureaux-de-vote-en-2012/?tab=table)
}
\usage{
ParisPollingStations2012
}
\description{
A shapefile of the polling stations in Paris in 2012.
}
\keyword{datasets}

