% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RP_2011_CS8_Paris}
\alias{RP_2011_CS8_Paris}
\title{Socioprofessionnal making of the population over 18 in Paris}
\format{An object of class \code{data.frame} with 992 rows and 10 columns.}
\source{
[INSEE](http://www.insee.fr/fr/themes/detail.asp?reg_id=99&ref_id=infra-population-11)
}
\usage{
RP_2011_CS8_Paris
}
\description{
A dataset containing the number of inhabitants by socioprofessional category in 2011 in Paris, by IRIS (the French census unit).
}
\keyword{datasets}

