% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spReapportion.R
\name{spReapportion}
\alias{spReapportion}
\title{Reapportion data from one geography to another}
\usage{
spReapportion(old_geom, new_geom, data, old_ID, new_ID, data_ID, variables)
}
\arguments{
\item{old_geom}{a \code{SpatialPolygonsDataFrame} representing the initial geometry.}

\item{new_geom}{a \code{SpatialPolygonsDataFrame} representing the geometry you want to reapportion data to.}

\item{data}{a \code{data.frame} containing the data to reapportion, and an ID allowing to match it to the \code{old_geom}.}

\item{old_ID}{a string, the name of the ID variable in the \code{old_geom}.}

\item{new_ID}{a string, the name of the ID variable in the \code{new_geom}.}

\item{data_ID}{a string, the name of the ID variable in the \code{data}.}

\item{variables}{a character vector, representing the names of the variables in the \code{data} set to reapportion. By default, all data variables except for the ID.}
}
\description{
This function allows to reapportion data from one geography to another, for example in the context of working with different administrative units.
}
\note{
Inspiration from http://stackoverflow.com/a/17703903 and http://rstudio-pubs-static.s3.amazonaws.com/6577_3b66f8d8f4984fb2807e91224defa854.html. All mistakes are mine, obviously.
}
\author{
Joel Gombin
}

