% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RP_2011_CS8_Paris}
\alias{RP_2011_CS8_Paris}
\title{Socioprofessionnal making of the population over 18 in Paris}
\format{\preformatted{'data.frame':	992 obs. of  10 variables:
 $ IRIS          : chr  "751010101" "751010102" "751010103" "751010104" ...
 $ C11_POP15P    : int  784 171 415 11 0 0 2157 1576 2115 2174 ...
 $ C11_POP15P_CS1: int  0 0 0 0 0 0 4 0 8 3 ...
 $ C11_POP15P_CS2: int  45 15 29 0 0 0 92 72 110 80 ...
 $ C11_POP15P_CS3: int  294 38 147 7 0 0 864 628 757 748 ...
 $ C11_POP15P_CS4: int  97 17 55 3 0 0 357 188 323 322 ...
 $ C11_POP15P_CS5: int  68 6 22 0 0 0 240 181 233 324 ...
 $ C11_POP15P_CS6: int  13 16 10 0 0 0 55 34 68 90 ...
 $ C11_POP15P_CS7: int  136 55 53 0 0 0 272 256 333 304 ...
 $ C11_POP15P_CS8: int  132 24 99 1 0 0 274 217 282 302 ...
}}
\source{
\href{http://www.insee.fr/fr/themes/detail.asp?reg_id=99&ref_id=infra-population-11}{INSEE}
}
\usage{
RP_2011_CS8_Paris
}
\description{
A dataset containing the number of inhabitants by socioprofessional category in 2011 in Paris, by IRIS (the French census unit).
}
\keyword{datasets}

