% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RP_2011_CS8_Paris}
\alias{RP_2011_CS8_Paris}
\title{Socioprofessionnal making of the population over 18 in Paris}
\format{\preformatted{'data.frame':	992 obs. of  10 variables:
 $ IRIS          : chr  "751010101" "751010102" "751010103" "751010104" ...
 $ C11_POP15P    : int  784 171 415 11 0 0 2157 1576 2115 2174 ...
 $ C11_POP15P_CS1: num  0 0 0 0 NaN ...
 $ C11_POP15P_CS2: num  5.74 8.77 6.99 0 NaN ...
 $ C11_POP15P_CS3: num  37.5 22.2 35.4 63.6 NaN ...
 $ C11_POP15P_CS4: num  12.37 9.94 13.25 27.27 NaN ...
 $ C11_POP15P_CS5: num  8.67 3.51 5.3 0 NaN ...
 $ C11_POP15P_CS6: num  1.66 9.36 2.41 0 NaN ...
 $ C11_POP15P_CS7: num  17.3 32.2 12.8 0 NaN ...
 $ C11_POP15P_CS8: num  16.84 14.04 23.86 9.09 NaN ...
}}
\source{
\href{http://www.insee.fr/fr/themes/detail.asp?reg_id=99&ref_id=infra-population-11}{INSEE}
}
\usage{
RP_2011_CS8_Paris
}
\description{
A dataset containing the number of inhabitants by socioprofessional category in 2011 in Paris, by IRIS (the French census unit).
}
\keyword{datasets}

