% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spReapportion.R
\name{spReapportion}
\alias{spReapportion}
\title{Reapportion data from one geography to another}
\usage{
spReapportion(old_geom, new_geom, data, old_ID, new_ID, data_ID,
  variables = names(data)[-which(names(data) \%in\% data_ID)],
  mode = "count", weights = NULL, weight_matrix = NULL,
  weight_matrix_var = NULL)
}
\arguments{
\item{old_geom}{a `SpatialPolygonsDataFrame` representing the initial geometry.}

\item{new_geom}{a `SpatialPolygonsDataFrame` representing the geometry you want to reapportion data to.}

\item{data}{a `data.frame` containing the data to reapportion, and an ID allowing to match it to the `old_geom`.}

\item{old_ID}{a string, the name of the ID variable in the `old_geom`.}

\item{new_ID}{a string, the name of the ID variable in the `new_geom`.}

\item{data_ID}{a string, the name of the ID variable in the `data`.}

\item{variables}{a character vector, representing the names of the variables in the `data` set to reapportion. By default, all data variables except for the ID.}

\item{mode}{either `"count"` or `"proportion"`. `"count"` is for absolute values, `"proportion"` is for, well, proportions (expressed between 0 and 1). If `"proportion"`, you need to provide a weights variable.}

\item{weights}{In case the variables are proportions, the name of the variable containing weights (i.e. the total number of observations per unit in the `old_geom`).}

\item{weight_matrix}{a SpatialPointsDataFrame indicating where are the observations (inhabitants, voters, etc.) (optional).}

\item{weight_matrix_var}{the name of the variable in \code{weight_matrix} containing the weights.}
}
\description{
This function allows to reapportion data from one geography to another, for example in the context of working with different administrative units.
}
\note{
Inspiration from http://stackoverflow.com/a/17703903 and http://rstudio-pubs-static.s3.amazonaws.com/6577_3b66f8d8f4984fb2807e91224defa854.html. All mistakes are mine, obviously.
}
\author{
Joel Gombin
}

